Factor out Visualizer from Learner
This lets us visualize data samples directly without creating a Learner
