+++ rastervision_pytorch_learner/rastervision/pytorch_learner/dataset/__init__.py
@@ -7,3 +7,4 @@ from rastervision.pytorch_learner.dataset.classification_dataset import *
 from rastervision.pytorch_learner.dataset.object_detection_dataset import *
 from rastervision.pytorch_learner.dataset.semantic_segmentation_dataset import *
 from rastervision.pytorch_learner.dataset.regression_dataset import *
+from rastervision.pytorch_learner.dataset.visualizer import *
