+++ rastervision_pytorch_learner/rastervision/pytorch_learner/regression_learner.py
@@ -1,19 +1,17 @@
-from typing import TYPE_CHECKING, Optional, Sequence
+from typing import TYPE_CHECKING, Optional
 import warnings
 from os.path import join
 import logging
 
 import matplotlib.pyplot as plt
 import matplotlib.gridspec as gridspec
-from textwrap import wrap
 
-import numpy as np
 import torch
 import torch.nn.functional as F
 
 from rastervision.pytorch_learner.learner import Learner
-from rastervision.pytorch_learner.utils.utils import (plot_channel_groups,
-                                                      channel_groups_to_imgs)
+from rastervision.pytorch_learner.dataset.visualizer import (
+    RegressionVisualizer)
 
 if TYPE_CHECKING:
     import torch.nn as nn
@@ -24,6 +22,9 @@ log = logging.getLogger(__name__)
 
 
 class RegressionLearner(Learner):
+    def get_visualizer_class(self):
+        return RegressionVisualizer
+
     def build_model(self, model_def_path: Optional[str] = None) -> 'nn.Module':
         """Override to pass class_names, pos_class_names, and prob_class_names.
         """
@@ -86,67 +87,6 @@ class RegressionLearner(Learner):
     def prob_to_pred(self, x):
         return x
 
-    def get_plot_ncols(self, **kwargs) -> int:
-        ncols = len(self.cfg.data.plot_options.channel_display_groups) + 1
-        return ncols
-
-    def plot_xyz(self,
-                 axs: Sequence,
-                 x: torch.Tensor,
-                 y: int,
-                 z: Optional[int] = None) -> None:
-
-        channel_groups = self.cfg.data.plot_options.channel_display_groups
-
-        img_axes = axs[:-1]
-        label_ax = axs[-1]
-
-        # plot image
-        imgs = channel_groups_to_imgs(x, channel_groups)
-        plot_channel_groups(img_axes, imgs, channel_groups)
-
-        # plot label
-        class_names = self.cfg.data.class_names
-        class_names = ['-\n-'.join(wrap(c, width=8)) for c in class_names]
-        if z is None:
-            # display targets as a horizontal bar plot
-            bars_gt = label_ax.barh(
-                y=class_names, width=y, color='lightgray', edgecolor='black')
-            # show values on the end of bars
-            label_ax.bar_label(bars_gt, fmt='%.3f', padding=3)
-
-            label_ax.set_title('Ground truth')
-        else:
-            # display targets and predictions as a grouped horizontal bar plot
-            bar_thickness = 0.35
-            y_tick_locs = np.arange(len(class_names))
-            bars_gt = label_ax.barh(
-                y=y_tick_locs + bar_thickness / 2,
-                width=y,
-                height=bar_thickness,
-                color='lightgray',
-                edgecolor='black',
-                label='true')
-            bars_pred = label_ax.barh(
-                y=y_tick_locs - bar_thickness / 2,
-                width=z,
-                height=bar_thickness,
-                color=plt.get_cmap('tab10')(0),
-                edgecolor='black',
-                label='pred')
-            # show values on the end of bars
-            label_ax.bar_label(bars_gt, fmt='%.3f', padding=3)
-            label_ax.bar_label(bars_pred, fmt='%.3f', padding=3)
-
-            label_ax.set_yticks(ticks=y_tick_locs, labels=class_names)
-            label_ax.legend(
-                ncol=2, loc='lower center', bbox_to_anchor=(0.5, 1.0))
-
-        label_ax.xaxis.grid(linestyle='--', alpha=1)
-        label_ax.set_xlabel('Target value')
-        label_ax.spines['right'].set_visible(False)
-        label_ax.get_yaxis().tick_left()
-
     def eval_model(self, split):
         super().eval_model(split)
 
@@ -180,7 +120,6 @@ class RegressionLearner(Learner):
             ax.set_ylabel('predictions')
         scatter_path = join(self.output_dir, '{}_scatter.png'.format(split))
         plt.savefig(scatter_path)
-        print('done scatter')
 
         # make histogram of errors
         fig = plt.figure(
@@ -196,4 +135,3 @@ class RegressionLearner(Learner):
             ax.set_xlabel('prediction error')
         hist_path = join(self.output_dir, '{}_err_hist.png'.format(split))
         plt.savefig(hist_path)
-        print('done hist')
